+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## v0.2.0 / 2017 Nov 13
+
+> This release adds basic email sending. We take the subject and body submitted
+> on the contact form and send it off in an email.
+
+* **Add** - email the data submitted via `POST /v1/contact`
+
+```clojure
+[ilyab "0.2.0"]
+```
+
+
 ## v0.1.0 / 2017 Nov 11
 
 > This is the initial release of the `ilyab` website and service.